// face-check-ui/src/app/services/models/page-response-attendance-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AttendanceResponse } from '../models/attendance-response';
export interface PageResponseAttendanceResponse {
  content?: Array<AttendanceResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElement?: number;
  totalPages?: number;
}