// face-check-ui/src/app/services/models/attendance-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AttendanceResponse {
  attendanceId?: number;
  checkInLatitude?: number;
  checkInLocation?: string;
  checkInLongitude?: number;
  checkInPhotoUrl?: string;
  checkInTime?: string;
  checkOutLatitude?: number;
  checkOutLocation?: string;
  checkOutLongitude?: number;
  checkOutPhotoUrl?: string;
  checkOutTime?: string;
  companyId?: number;
  companyName?: string;
  email?: string;
  firstName?: string;
  grossPayPerDay?: number;
  hoursWorked?: number;
  lastName?: string;
  netPay?: number;
  overtimeHours?: number;
  periodEnd?: string;
  periodStart?: string;
  phone?: string;
  workerId?: number;
}