// face-check-ui/src/app/services/models/overtime-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface OvertimeResponse {
  approvedBy?: number;
  attendanceId?: number;
  date?: string;
  isSuccessful?: boolean;
  message?: string;
  overtimeHours?: number;
  reason?: string;
  regularHours?: number;
  workerId?: number;
  workerName?: string;
}